Refactor fgetline(), chop(), partial hex2bin() usage
