+++ hcxhash2cap.c
@@ -25,6 +25,7 @@
 #include "include/ieee80211.c"
 #include "include/strings.c"
 #include "include/byteops.c"
+#include "include/fileops.c"
 
 #define ARCH_INDEX(x)	((unsigned int)(unsigned char)(x))
 
@@ -489,42 +490,6 @@ if(write(fd_cap, packetout, PCAPREC_SIZE +MAC_SIZE_NORM +CAPABILITIESAP_SIZE +2 
 return;
 }
 /*===========================================================================*/
-static size_t chop(char *buffer, size_t len)
-{
-static char *ptr;
-
-ptr = buffer +len -1;
-while(len)
-	{
-	if (*ptr != '\n')
-		break;
-	*ptr-- = 0;
-	len--;
-	}
-while(len)
-	{
-	if (*ptr != '\r')
-		break;
-	*ptr-- = 0;
-	len--;
-	}
-return len;
-}
-/*---------------------------------------------------------------------------*/
-static int fgetline(FILE *inputstream, size_t size, char *buffer)
-{
-static size_t len;
-static char *buffptr;
-
-if(feof(inputstream))
-	return -1;
-buffptr = fgets (buffer, size, inputstream);
-if(buffptr == NULL)
-	return -1;
-len = strlen(buffptr);
-len = chop(buffptr, len);
-return len;
-}
 static uint16_t getfield(char *lineptr, size_t bufflen, uint8_t *buff)
 {
 static size_t p;
