+++ include/fileops.h
@@ -1,3 +1,5 @@
+#define ESSID_LEN_MAX 32
+
 /*===========================================================================*/
 struct magicnr_s
 {
