+++ whoismac.c
@@ -13,6 +13,7 @@
 #include <curl/curl.h>
 
 #include "include/strings.c"
+#include "include/fileops.c"
 
 #define LINEBUFFER_MAX	256
 #define OUIBUFFER_MAX	8192
@@ -62,38 +63,6 @@ while (!feof(fhouitmp))
 fclose(fhoui);
 fprintf(stdout, "\ndownload finished\n");
 return true;
-}
-/*===========================================================================*/
-static size_t chop(char *buffer,  size_t len)
-{
-static char *ptr;
-
-ptr = buffer +len -1;
-while (len) {
-	if (*ptr != '\n') break;
-	*ptr-- = 0;
-	len--;
-	}
-
-while (len) {
-	if (*ptr != '\r') break;
-	*ptr-- = 0;
-	len--;
-	}
-return len;
-}
-/*---------------------------------------------------------------------------*/
-static int fgetline(FILE *inputstream, size_t size, char *buffer)
-{
-if (feof(inputstream)) return -1;
-		char *buffptr = fgets (buffer, size, inputstream);
-
-	if (buffptr == NULL) return -1;
-
-	size_t len = strlen(buffptr);
-	len = chop(buffptr, len);
-
-return len;
 }
 /*===========================================================================*/
 static void getessidinfo(char *essidname)
